fix: import types from @rspack/core
