+++ package.json
@@ -53,6 +53,8 @@
     "@babel/core": "^7.7.7",
     "@babel/preset-env": "^7.11.5",
     "@intlify/vue-i18n-loader": "^3.0.0",
+    "@rspack/core": "^1.7.5",
+    "@rstest/core": "^0.7.9",
     "@types/cssesc": "^3.0.2",
     "@types/estree": "^0.0.45",
     "@types/jsdom": "^16.2.13",
@@ -66,7 +68,6 @@
     "file-loader": "^6.1.0",
     "html-webpack-plugin": "^4.5.0",
     "html-webpack-plugin-v5": "npm:html-webpack-plugin@^5.3.2",
-    "@rstest/core": "^0.7.9",
     "jsdom": "^16.4.0",
     "lint-staged": "^10.3.0",
     "markdown-loader": "^6.0.0",
@@ -85,7 +86,7 @@
     "sugarss": "^3.0.1",
     "ts-loader": "^8.0.6",
     "ts-loader-v9": "npm:ts-loader@^9.2.4",
-    "typescript": "^4.4.3",
+    "typescript": "^5.9.3",
     "url-loader": "^4.1.0",
     "vue": "^3.4.3",
     "vue-i18n": "^9.1.7",
