+++ src/resolveScript.ts
@@ -1,4 +1,4 @@
-import type { LoaderContext } from 'webpack'
+import type { LoaderContext } from '@rspack/core'
 import type {
   SFCDescriptor,
   SFCScriptBlock,
