+++ src/select.ts
@@ -1,4 +1,4 @@
-import type { LoaderContext } from 'webpack'
+import type { LoaderContext } from '@rspack/core'
 import type { SFCDescriptor } from 'vue/compiler-sfc'
 import type { ParsedUrlQuery } from 'querystring'
 import { resolveScript } from './resolveScript'
